Issue DP-212 - Fixed crashes in Chromium integration
Added AdblockHelper.deinit() to use that in Chromium when
FilterEngine is disposed.

Added initPreferences() also to onCreate() in
GeneralSettingsFragment to avoid crash when this fragment
is being restored from background and it's child views
are already displayed (for example Filters List).
Otherwise we could hit crashes like "MultiSelectListPreference
requires an entries array and an entryValues array." thrown
from performLaunchActivity().

Updated README.
Bump versions to 3.8.1.

Signed-off-by: Krystian Zlomek <k.zlomek@eyeo.com>