Issue DP-588 - Video player on Wikihow doesn't play...
Problem was introduced (as a regression) by the site key feature.
Something is wrong when we are making custom requests using
a "Range" header, for example on the testing page -
https://m.wikihow.com/Video/Use-a-Kick-Board - video player won't
play unless user clicks serveral time Play button.
No error is thrown or printed and responses and requests look ok.

As it is unclear what is wrong and how/if it can be fixed while
site key check for the requests which contain "Range" header
(which should be used for media files) does not seem to be a valid
use case for the site key feature in general, we are removing here
the site key check for all requests which include "Range" header.

Signed-off-by: Krystian Zlomek <k.zlomek@eyeo.com>