Issue DP-703 - Add publish to bintray CI job
This adds a new step to the pipeline which automatically publishes
artifacts to bintray when the build runs on a tagged checkout of the code.

The credentials and tagging mechanism are both protected which should stop any credentials leaks.

By default the job runs with the DRYRUN=true.
To actually publish something you must kick off a manual build and set this to false