Noissue - Make AndroidHttpClientTests faster
This change usually reduces tests time from ~11 sec to ~2 sec
on my device.

But sometimes without this change tests run for ~2 minutes.

This is because every time we spawn FE with a normal AndroidHttpClient
we download both FL files which takes some of time.

This change blocks FL downloads and filters are not needed here.

Signed-off-by: Krystian Zlomek <k.zlomek@eyeo.com>