Issue DP-1073 - Return InputStream instead of a ByteBuffer
This change will allow WebView internals to manage response
data, which should be done in a more efficient way than we
did it so far.

For exmaple when loading http://incident.net/v8/files/mp4/5.mp4
without this chnage WebView waits for ~7 seconds for HttpClient
to finish reading data instead of returning an input stream.

Signed-off-by: Krystian Zlomek <k.zlomek@eyeo.com>