Issue DP-1219 - Restore old look of the filters list UI.
Previously in the "specialization" property a string to be presented
in filters selection UI was present. This however got removed when
moving to core 3.6. Core does not provide any equivalent. It seems this
now is left as UI's responsibility. This is why the extension way was
followed i.e. having a locate->native language map and using
subscription's language in order to get the native language
and presenting that.