DP-1393 - Set correct input stream for redirections
When redirection happens (and is not handled automatically by
our AndroidHttpClient based on request.getFollowRedirect() == false)
then connection.getErrorStream() is null while
connection.getInputStream() is not null but empty (no bytes to read).

We still need to propagate this connection.getInputStream()
to the caller as otherwise underlying WebView will trigger the
request once again.

Signed-off-by: Krystian Zlomek <k.zlomek@eyeo.com>