DP-1449 - Updated how we set a top level referrer
This changes a bit how we add a top level url:
now we will compare domains not only whole url
before we decide if we should add a top level
parent.

This is to avoid the problem when FE matcher may
return a different result when a url/domain is
added twice on the referrers list.

Note that this change does not only affect EH but
also network blocking/whitelisting.

Signed-off-by: Krystian Zlomek <k.zlomek@eyeo.com>